add java-time lib and calculate payday for each installment.
To keep thing simple, for now the default days between each installment
is a commercial month, 30 days.
Using that premise we ensure that the interest in each parcel is calculated
correctly. Since a longer/shorter period would change the interest for that
installment.
And for now we must assume that the loan start date is the date of the simulation